pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bss-rebalancing"
version = "1.0.0"
description = "Fully Dynamic Rebalancing of Dockless Bike Sharing Systems using Deep Reinforcement Learning"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "CC-BY-NC-4.0"}
authors = [
    {name = "Edoardo Scarpel", email = "edoardo.scarpel@studenti.unipd.it"}
]
keywords = [
    "bike-sharing",
    "reinforcement-learning",
    "deep-learning",
    "urban-mobility",
    "rebalancing",
    "deep-q-network",
    "gymnasium"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: GIS",
]

dependencies = [
    # This will install all subpackages
    "bss-preprocessing",
    "bss-gymnasium-env",
    "bss-rl-training",
    "bss-benchmark",
    "bss-results-webapp",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/edos08/BSS-DynamicRebalancing-RL"
Repository = "https://github.com/edos08/BSS-DynamicRebalancing-RL"
Thesis = "https://hdl.handle.net/20.500.12608/84368"
Issues = "https://github.com/edos08/BSS-DynamicRebalancing-RL/issues"

[tool.setuptools]
# Install all subpackages when installing the root package
packages = {find = {where = ["src/bss_rebalancing/**/src"]}}

[tool.setuptools.package-data]
"*" = ["*.yml", "*.yaml", "*.json"]

# When you run 'pip install -e .' at root, it will:
# 1. Discover all subpackages in src/bss_rebalancing/*/
# 2. Install them in editable mode
# 3. Register all CLI entry points from their pyproject.toml files
